+++ static/components/f-file-list.js
@@ -1,4 +1,4 @@
-import { getFieldName, initField } from "./utils.js";
+import { getFieldName, getItemLabel, initField } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -32,11 +32,9 @@ customElements.define(
 
       const div = dom("div", { class: "fieldset" }, this);
       let index = 0;
-      const firstKey = schema.fields[0].name;
 
       function createOption(value, isNew = false) {
-        const label = (typeof value === "object" && value[firstKey]) ||
-          "New item";
+        const label = getItemLabel(schema, value, index);
         const attributes = schema.attributes || {};
         attributes.open ??= isNew;
 
@@ -117,13 +115,11 @@ customElements.define(
       this.querySelector(".field-description").innerHTML = schema.description ??
         "";
       const items = Array.from(this.querySelector(".fieldset").children);
-      const firstKey = schema.fields[0].name;
 
       for (const value of values) {
-        const label = value[firstKey];
         items.shift()?.update({
           type: "object",
-          label,
+          label: getItemLabel(schema, value),
           fields: schema.fields,
         }, value);
       }